+++ afat/helper/views.py
@@ -23,12 +23,11 @@ from allianceauth.framework.api.user import get_main_character_name_from_user
 from allianceauth.services.hooks import get_extension_logger
 
 # Alliance Auth AFAT
-from afat import __title__
 from afat.helper.users import users_with_permission
 from afat.models import Fat, FatLink
-from afat.providers import AppLogger
+from afat.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 
 class AFATUI(Enum):