+++ afat/utils.py
@@ -14,11 +14,10 @@ from allianceauth.eveonline.models import (
 from allianceauth.services.hooks import get_extension_logger
 
 # Alliance Auth AFAT
-from afat import __title__
-from afat.handler import esi_handler
-from afat.providers import AppLogger, esi
+from afat.providers.applogger import AppLogger
+from afat.providers.esi import ESIHandler, esi
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 # Format for output of datetime for this app
 DATETIME_FORMAT = "%Y-%m-%d %H:%M"
@@ -82,7 +81,7 @@ def get_or_create_character(
     if name:
         # If a name is passed to this function, we have to check it on ESI
         operation = esi.client.Universe.PostUniverseIds(body=[name])
-        result = esi_handler.result(operation, use_etag=False)
+        result = ESIHandler.result(operation, use_etag=False)
 
         if not result or not result.characters:
             return None